+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   type.
 - New #125: Add `editUserStarSubscription` method.
 - New #125: Add `subscriptionPeriod` field to `TransactionPartnerUser` type.
+- New #125: Add `setUserEmojiStatus` method.
 
 ## 0.4.5 October 31, 2024